Time System Working With Sleeping
The time now works and when a set time is reached the player goes to sleep and is teleported to the bed. A UI has been added to see the current time and date in the top right. This is a place holder and will be depreciated and time will be added to the main UI.

An error is thrown every time the player is teleported because the character controller is disabled and it is set to check for that. This will be disabled in future builds to keep the console clean.

Current Issues to fix.

Time not formatted for AM & PM. Time will increment every hour and will never turn over from 12 -> 1 or 24 -> 0